% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read in FARS File}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the name of the FARS file.}
}
\value{
This function returns a tibble of the data contained in the
file named \code{filename} and prints the first 10 rows of the
tibble.
}
\description{
This is a function which checks whether a FARS file with a certain
file name exists.
An error will be thrown if a file by the user-supplied file name does
not exist.
If said file does exist, it will be read into R and transformed into
a tibble.
}
\examples{
\dontrun{fars_read("accident_2014.csv.bz2")}
\dontrun{fars_read("accident_2013.csv.bz2")}

}
