% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read in files by year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list or vector of years.}
}
\value{
A list of tibbles, each with two columns: one containing the
months, and the other containing the year.
The number of rows for each month represents the number of
accidents in the month.
This list is also printed.
}
\description{
This function takes a list of years and creates file names for each of
them using the \code{make_filename} function.
It then reads in these files using the \code{fars_read} function and
creates a list of tibbles which contain the months in each of the
years.
If the file name created from a year inputted does not correspond
to an existing file containing data for said year, an error is thrown.
}
\examples{
\dontrun{fars_read_years(c(2014,2015))}
\dontrun{fars_read_years(c(2013,2014))}

}
