% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize yearly FARS data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list or vector of years.}
}
\value{
A tibble containing the number of accidents in each month
for each of the years specified in the list of years supplied to
the \code{years} argument.
}
\description{
This function passes a list of years to the \code{fars_read_years}
function. It then counts the number of rows for each month in the
data frame returned by \code{fars_read_years} to obtain the number of
accidents that occurred in each month for the years supplied to the
\code{years} argument.
It then returns a tibble containing this information.
}
\examples{
\dontrun{fars_summarize_years(c(2013,2014,2015))}
\dontrun{fars_summarize_years(c(2013,2014))}

}
