% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{File Naming Function}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A character string giving the year for which a file name
should be generated.}
}
\value{
A character string giving a file name formatted
\code{accident_X.csv.bz2} where X is the string supplied to
\code{year}.
}
\description{
This function makes a FARS file name for a given year. It converts the
string supplied by the user to an integer, then generates a file name
for that year with the correct format.
An error will be thrown if the supplied string cannot be coerced to
an integer without introducing NA's.
}
\examples{
\dontrun{make_filename("2010")}
\dontrun{make_filename("2021")}

}
